Allow for Modification of Default HTTP Response Headers (#321)
* Write tests for default HTTP response headers mods

* Allow over-riding of default HTTP response headers via configuration

* Allow custom default headers to be specified via CLI arguments

* Apply changes introduced by black linter

* Remove test for over-riding date header

@tomchristie requested to not worry about over-riding the date header,
so there is no need for this test anymore.

* Rename kwarg/cli arguments and move header logic from Server to Config

As requested in PR #321, I'm renaming the --default-header CLI argument
to --header, and renaming the kwarg/instance variable custom_headers to
headers. Also, I'm moving (most) of the header generation logic from
main.Server to config.Config, so that most of the headers only get
generated once, instead of on every tick. The most computationally
intense generation will still happen ever tick, and that is the
date-stamp generation.

* Rename instance fields relating to custom default headers